COM-2752 rename with plural
