+++ src/helpers/vendorCompanies.js
@@ -1,0 +1,3 @@
+const VendorCompany = require('../models/VendorCompany');
+
+exports.get = async () => VendorCompany.findOne().lean();
