+++ src/models/VendorCompany.js
@@ -1,10 +1,5 @@
 const mongoose = require('mongoose');
-const {
-  validateQuery,
-  validateUpdateOne,
-  formatQuery,
-  formatQueryMiddlewareList,
-} = require('./preHooks/validate');
+const { formatQuery, formatQueryMiddlewareList } = require('./preHooks/validate');
 const { SIRET_VALIDATION } = require('./utils');
 const addressSchemaDefinition = require('./schemaDefinitions/address');
 
@@ -12,11 +7,8 @@ const VendorCompanySchema = mongoose.Schema({
   name: { type: String, required: true, unique: true },
   address: { type: mongoose.Schema(addressSchemaDefinition, { _id: false, id: false }), required: true },
   siret: { type: String, validate: SIRET_VALIDATION, required: true, unique: true },
-}, { timestamps: true, collection: 'vendorcompany' });
+}, { timestamps: true });
 
-VendorCompanySchema.pre('findOne', validateQuery);
-VendorCompanySchema.pre('updateOne', validateUpdateOne);
 formatQueryMiddlewareList().map(middleware => VendorCompanySchema.pre(middleware, formatQuery));
 
 module.exports = mongoose.model('VendorCompany', VendorCompanySchema);
-module.exports.SIRET_VALIDATION = SIRET_VALIDATION;
